requirements.txt
django
psycopg2-binary
Pillow
django-ckeditor
